Update u32 -> ulong for the entirety of nw4r

+++ include/lib/nw4r/ef/drawstrategy/ef_drawstrategybuilder.h
@@ -17,7 +17,7 @@ enum DrawStrategyType {
 
 class DrawStrategyBuilder {
 public:
-    virtual DrawStrategy* Create(u32 type); // at 0x8
+    virtual DrawStrategy* Create(ulong type); // at 0x8
 };
 
 } // namespace ef
