+++ include/lib/nw4r/g3d/g3d_anmclr.h
@@ -23,7 +23,7 @@ class AnmObjMatClrRes;
 class AnmObjMatClr : public AnmObj {
 public:
     AnmObjMatClr(MEMAllocator* pAllocator, u16* pBindingBuf, int numBinding);
-    virtual void G3dProc(u32 task, u32 param, void* pInfo) = 0; // at 0xC
+    virtual void G3dProc(ulong task, ulong param, void* pInfo) = 0; // at 0xC
     virtual ~AnmObjMatClr() {}                                  // at 0x10
 
     virtual void SetFrame(f32 frame) = 0; // at 0x1C
@@ -37,14 +37,14 @@ public:
     virtual void Release();                  // at 0x34
 
     virtual const ClrAnmResult* GetResult(ClrAnmResult* pResult,
-                                          u32 idx) = 0; // at 0x38
+                                          ulong idx) = 0; // at 0x38
 
     virtual AnmObjMatClrRes* Attach(int idx, AnmObjMatClrRes* pRes); // at 0x3C
     virtual AnmObjMatClrRes* Detach(int idx);                        // at 0x40
     virtual void DetachAll();                                        // at 0x44
 
-    bool TestExistence(u32 idx) const;
-    bool TestDefined(u32 idx) const;
+    bool TestExistence(ulong idx) const;
+    bool TestDefined(ulong idx) const;
 
 protected:
     enum BindingFlag {
@@ -72,7 +72,7 @@ class AnmObjMatClrNode : public AnmObjMatClr {
 public:
     AnmObjMatClrNode(MEMAllocator* pAllocator, u16* pBindingBuf, int numBinding,
                      AnmObjMatClrRes** ppChildrenBuf, int numChildren);
-    virtual void G3dProc(u32 task, u32 param, void* pInfo); // at 0xC
+    virtual void G3dProc(ulong task, ulong param, void* pInfo); // at 0xC
     virtual ~AnmObjMatClrNode();                            // at 0x10
 
     virtual void SetFrame(f32 frame); // at 0x1C
@@ -118,7 +118,7 @@ public:
     virtual ~AnmObjMatClrOverride() {} // at 0x10
 
     virtual const ClrAnmResult* GetResult(ClrAnmResult* pResult,
-                                          u32 idx); // at 0x38
+                                          ulong idx); // at 0x38
 
     NW4R_G3D_RTTI_DECL_DERIVED(AnmObjMatClrOverride, AnmObjMatClrNode);
 };
@@ -135,7 +135,7 @@ public:
 
     AnmObjMatClrRes(MEMAllocator* pAllocator, ResAnmClr clr, u16* pBindingBuf,
                     int numBinding, ClrAnmResult* pCacheBuf);
-    virtual void G3dProc(u32 task, u32 param, void* pInfo); // at 0xC
+    virtual void G3dProc(ulong task, ulong param, void* pInfo); // at 0xC
     virtual ~AnmObjMatClrRes() {}                           // at 0x10
 
     virtual void SetFrame(f32 frame); // at 0x1C
@@ -148,7 +148,7 @@ public:
     virtual bool Bind(const ResMdl mdl); // at 0x30
 
     virtual const ClrAnmResult* GetResult(ClrAnmResult* pResult,
-                                          u32 idx); // at 0x38
+                                          ulong idx); // at 0x38
 
     void UpdateCache();
